requirements.txt
asgiref==3.5.2
Django==4.1.2
gunicorn==20.1.0
sqlparse==0.4.3
tzdata==2022.4
django-crispy-forms==1.14.0